backupmate/config.py
import os
from dotenv import load_dotenv

def load_config(env_path=".backupmate.env"):
    """Loads configuration parameters from the .env file."""
    load_dotenv(dotenv_path=env_path)
    # Placeholder for config loading logic
    print("Loading configuration (placeholder)")
    return {}

def validate_config(config):
    """Validates the loaded configuration."""
    # Placeholder for config validation logic
    print("Validating configuration (placeholder)")
    return True
