pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "backupmate"
version = "0.1.0"
dependencies = [
    "python-dotenv",
    "boto3",
    "argparse",
    "SQLAlchemy"
]
